% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gor-package.R
\docType{package}
\encoding{UTF-8}
\name{gor}
\alias{gor}
\title{Graphs and Network Optimization algorithms}
\description{
This is a hardly complete collection of algorithms
from the subject "Graphs and Network Optimization".
}
\details{
Functions in this package have been written for teaching
purposes.  Therefore, no attempt at production versions of the
algorithms has been made.  They are neither complete nor
completely correct, although a great effort has been invested
in their construction and debugging.  All of them pass a
series of tests given in the examples sections of the
corresponding help pages.  Comments and suggestions are
welcome, even if I cannot guarantee that they will be
incorporated to the package.

This package makes extensive use of the \link{igraph} package
functions, which should be loaded before using "gor".

Some functions in this package perform tasks which can be found in
other well-tested packages such as \link{igraph} or TSP.  As said
before, these functions have been written with teaching in
mind, and I do not claim that \link{gor} functions are better than
any other in any way whatsoever.
}
\author{
Cesar Asensio (2021-2023)
}
