% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-grafos-TSP.R
\encoding{UTF-8}
\name{next_index}
\alias{next_index}
\title{Next position to i in a cycle}
\usage{
next_index(i, n)
}
\arguments{
\item{i}{Position in cycle}

\item{n}{Lenght of cycle}
}
\value{
The next position in cycle
}
\description{
Next position to i in a cycle.
}
\details{
In a cycle, the next slot to the i-th position is i+1 unless i=n.
In this case, the next is 1.
}
\examples{
next_index(5, 7)   # 6
next_index(7, 7)   # 1

}
\author{
Cesar Asensio
}
