% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-grafos-NPdif.R
\encoding{UTF-8}
\name{sum_g}
\alias{sum_g}
\title{Sum of the higher terms of a list}
\usage{
sum_g(L, m)
}
\arguments{
\item{L}{A numeric sequence}

\item{m}{A scalar}
}
\value{
The sum of the m higher terms of the list L
}
\description{
Sum of the higher terms of a list
}
\details{
It sorts the list L in decreasing order and returns the sum of the
first m components of the ordered list.
}
\examples{
sum_g(1:10, 3)  # 8 + 9 + 10 = 27

}
\author{
Cesar Asensio
}
